Extract comment history panel
